// src/app/shared/models/reporte-cliente.ts
export class ReporteClienteCumpleanio{
  idCliente: number;
  nombre: string;
  edad: number;
  ultimaCita: Date | null;
  ultimaCitaColor: string | null;
  idUltimaCita: number | null;
  proximaCita: Date | null;
  proximaCitaColor: string | null;
  idProximaCita: number | null;
  constructor() {
  }
}